requirements.txt
PyYAML~=6.0.2
boto3~=1.35.58
botocore~=1.35.58